feat: * change optional data, now listing return always the least possible (core+variant), while for a single get it will be possible to choose what to get with granularity * Move to i64/f64 for every number type, wasting some memory but removing a lot of critical conversions between types * Create a build phase, used to initialize some sql table using data found in other DB's tables * Remove Creature 'from' bloated code. Now all the Creature structs must be initialized directly * remove creature info, squished the data in the core (roles) * remove create and scales cache. Creature could be reintroduced if latency is deemed too high * remove rawcreature, a long time coming. * add raw_query_builder.rs, helps building raw statement (no compile time checks) in a standardized way

+++ src/services/encounter_handler/difficulty_utilities.rs
@@ -1,6 +1,6 @@
 use crate::models::encounter_structs::{EncounterChallengeEnum, ExpRange};
 
-pub fn scale_difficulty_exp(base_difficulty: &EncounterChallengeEnum, party_size: i16) -> ExpRange {
+pub fn scale_difficulty_exp(base_difficulty: &EncounterChallengeEnum, party_size: i64) -> ExpRange {
     // Given the base difficulty and the party size, it scales the base difficulty.
     // Useful when a party is not the canon 4 party member.
     let party_deviation = party_size - 4;
@@ -17,7 +17,7 @@ pub fn scale_difficulty_exp(base_difficulty: &EncounterChallengeEnum, party_size
     }
 }
 
-fn convert_difficulty_enum_to_base_xp_budget(diff: &EncounterChallengeEnum) -> i16 {
+fn convert_difficulty_enum_to_base_xp_budget(diff: &EncounterChallengeEnum) -> i64 {
     match diff {
         EncounterChallengeEnum::Trivial => 40,
         EncounterChallengeEnum::Low => 60,
@@ -28,7 +28,7 @@ fn convert_difficulty_enum_to_base_xp_budget(diff: &EncounterChallengeEnum) -> i
     }
 }
 
-fn convert_difficulty_enum_to_xp_adjustment(diff: &EncounterChallengeEnum) -> i16 {
+fn convert_difficulty_enum_to_xp_adjustment(diff: &EncounterChallengeEnum) -> i64 {
     match diff {
         EncounterChallengeEnum::Trivial => 10,
         EncounterChallengeEnum::Low => 15,
